+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+
+## fixed
+
+- [ ] `Pfx.contrast/2` 1.1.1.0/24 is left-adjacent to 1.1.2.0/25, not disjoint!
+      Added left_nei and right_nei => left/right adjacent and combinable (neighors)
+       :left/:right means -> left/right adjacent, cannot be combined
+       :left_nei/:right_nei -> left/right adjacent, can be combined
+
 ## [0.10.1] - 2022-01-09
 
 ### fixed